Ignore consumable feature amount specified in derivations
Derivation can specify required features. Hydra supports consumable
features [1] by specifying an amount as suffix feature.
Basically, a derivation can specify a `requiredSystemFeature`
`feature:N` meaning this derivation requires `N` units of the feature
`feature`. The Hydra scheduler schedules jobs on Nix slaves by
respecting these constraints. The available amounts of features on
slaves are specified in the `machines` file.

Currently, before building a derivation, Nix checks that all required
features of this derivation are provided (they are specified via the
`system-features` option). Since the derivation needs `feature:N`,
`feature:N` must be specified in the `system-features` option
list. This is not convenient since the value of `N` can vary!

This patch proposes to ignore the suffix `:N` in the
`requiredSystemFeature` attribute of deviations. This allows Nix to
build derivations specifying consumable features by only specifying
the `feature` (and not `feature:N`) in the `system-features` list.

A warning is printed (`Chatty` mode) to inform the user consumable
feature amount is ignored.

    "warning: the amount of the consumable feature 'cpu:4' is ignored in derivation '/nix/store/ndf6b9r8wb6ljci3k1w3jddrsrscgm7d-vm-test-run-tcp-flows.drv'

Note this could break backward compatibility for users that are using
a `requiredFeatures` containing a `:`.

[1] https://github.com/NixOS/hydra/pull/588